requirements.txt
binance-connector==3.0.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
psycopg2==2.9.6
pycryptodome==3.18.0
pyTelegramBotAPI==4.12.0
requests==2.31.0
schedule==1.2.0
urllib3==2.0.2
websocket-client==1.5.2
